requirements.txt
fastapi
uvicorn
pydantic
python-jose
cryptography
pyjwt
pytest
requests
httpx


